+++ lib/main.dart
@@ -1,10 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:andre_house/UI/input_field.dart';
-import 'package:andre_house/StaffPages/staff_homepage.dart';
-import 'package:andre_house/GuestPages/guest_homepage.dart';
-
-void main() => runApp(const MyApp());
+import 'login_page.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
@@ -19,102 +16,3 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class HomeScreen extends StatefulWidget {
-  const HomeScreen({Key? key}) : super(key: key);
-  //StateFullWidget
-  @override
-  HomeScreenState createState() => HomeScreenState();
-}
-
-class HomeScreenState extends State<HomeScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      //Scaffold
-      body: Container(
-        width: MediaQuery.of(context).size.width,
-        height: MediaQuery.of(context).size.height,
-        color: Colors.lightBlueAccent,
-        child: Stack(
-          children: <Widget>[
-            Align(
-              alignment: Alignment.bottomRight,
-              heightFactor: 0.5,
-              widthFactor: 0.5,
-              child: Material(
-                borderRadius: const BorderRadius.all(Radius.circular(200.0)),
-                color: const Color.fromRGBO(255, 255, 255, 0.4),
-                child: Container(
-                  width: 400,
-                  height: 400,
-                ),
-              ),
-            ),
-            Center(
-              child: Container(
-                width: 400,
-                height: 400,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: <Widget>[
-                    Material(
-                        elevation: 10.0,
-                        borderRadius:
-                        const BorderRadius.all(Radius.circular(50.0)),
-                        child: Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: Image.asset(
-                            "images/flutter-logo.png",
-                            width: 80,
-                            height: 80,
-                          ),
-                        )),
-                    Form(
-                      child: InputField(
-                        //Calling inputField  class
-
-                          const Icon(
-                            Icons.person,
-                            color: Colors.white,
-                          ),
-                          "Username"),
-                    ),
-                    Form(
-                      child: InputField(
-                          const Icon(
-                            Icons.lock,
-                            color: Colors.white,
-                          ),
-                          "Password"),
-                    ),
-                    Container(
-                      width: 150,
-                      child: RaisedButton(
-                        //Raised Button
-                        onPressed: () {
-                          Navigator.of(context).push(MaterialPageRoute(
-                              builder: (context) => MyStaffHomePage()));
-                        },
-                        color: Colors.indigo,
-                        textColor: Colors.white,
-                        child: const Text(
-                          "Login",
-                          style: TextStyle(fontSize: 20.0),
-                        ),
-                        shape: const RoundedRectangleBorder(
-                          borderRadius: BorderRadius.all(
-                            Radius.circular(10.0),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
